repair: Move decorated_key_with_hash to separate header.
